Initial release of mediabackups software
Initial release based on the codebase from wmfbackups used
to generate the first full backups of mediabackups.

Many things here do not work as intended, are broken or untested,
but we want to generate an initial release to deploy and have
something usable on production in case an emergency recovery is
needed. Lots of work still needed to make this as good as
intended.

Bug: T276445
Bug: T300020
Change-Id: I4a616e9f07c546d8af2ac8870e81d272cbc580dd