Add functionality to "archiving" older status of a file
A new table, file_history is used to store older versions of the
file, so that a backup can be restored from an arbitrary point in
time.

The functionality has been added to the update script, so that
when it runs, when it detects the status of a file has changed,
it updates it, archiving the older status (file status, path,
timestamps) in the file_history table.

This will not detect renames/loss files yet- they will be stored
as new files (without storing them twice on backup storage).

This also assumes there are no duplicate files with the same title,
sha1 and upload timestamp- which is not true at the time.

WIP: T299764
Change-Id: Ic42aa0ae8ce995a9317373c624f4fbb4bd7ddd20