restore_media_file: Split restore-media-file cli into it and a library
In preparation to creating a delete-media-file and query-media-file
scripts, very similar to the restore ones, but for removing and
searching file, respectively, move common procedures to an interactive
query one, and use it from the existing restore-media-file script.

This should be a noop.

Further refactoring may be needed on the newly created library so
to better reuse the implemented procedures on the other scripts.

Bug: T311215
Change-Id: I8059a208c7863e18c5878ad1bffaf6e427dd2030