delete-media-file: Add failsafe to file deletion
In order to make sure we don't accidentally delete files- check
that the file is not publicly available before deletion.
Also allow for a deletion of several references of the same
backup.

Fixed at the same time an existing bug due to an earlier
refactoring when trying to retrieve foreign keys from the
database.

Bug: T311215
Change-Id: I89f0664b2ff259f363e96c65480f675411e19feb